dcp-815 separate class for metadata/resource.py
