+++ manifest/generator.py
@@ -5,7 +5,7 @@ from hca_ingest.api.ingestapi import IngestApi
 from exporter.graph.experiment_graph import ExperimentGraph
 from exporter.graph.graph_crawler import GraphCrawler
 from exporter.metadata.datafile import DataFile
-from exporter.scratch import MetadataResource
+from exporter.metadata.resource import MetadataResource
 from manifest.manifests import AssayManifest
