+++ tests/exporter/graph/test_graph_crawler.py
@@ -4,7 +4,8 @@ from hca_ingest.api.ingestapi import IngestApi
 from mock import MagicMock
 
 from exporter.graph.graph_crawler import GraphCrawler
-from exporter.scratch import MetadataResource, MetadataService
+from exporter.metadata.resource import MetadataResource
+from exporter.scratch import MetadataService
 from tests.mocks.files import MockEntityFiles
 from tests.mocks.ingest import MockIngestAPI
